+++ EntranceShuffle.py
@@ -5,7 +5,6 @@ from collections import defaultdict
 
 
 def link_entrances(world, player):
-    connect_two_way(world, 'Links House', 'Links House Exit', player) # unshuffled. For now
     connect_exit(world, 'Chris Houlihan Room Exit', 'Links House', player) # should always match link's house, except for plandos
 
     Dungeon_Exits = Dungeon_Exits_Base.copy()
@@ -80,7 +79,7 @@ def link_entrances(world, player):
 
         single_doors = list(Single_Cave_Doors)
         bomb_shop_doors = list(Bomb_Shop_Single_Cave_Doors)
-        blacksmith_doors = list(Blacksmith_Single_Cave_Doors)
+        blacksmith_doors = list(Blacksmith_Single_Cave_Doors) + ['Links House']
         door_targets = list(Single_Cave_Targets)
 
         # we shuffle all 2 entrance caves as pairs as a start
@@ -102,6 +101,18 @@ def link_entrances(world, player):
             exit1, exit2 = caves.pop()
             connect_two_way(world, entrance1, exit1, player)
             connect_two_way(world, entrance2, exit2, player)
+
+        # place links house
+        if world.mode[player] == 'standard' or not world.shufflelinks[player]:
+            links_house = 'Links House'
+        else:
+            links_house_doors = [i for i in LW_Single_Cave_Doors if i not in Isolated_LH_Doors_Open]
+            links_house = random.choice(links_house_doors)
+        connect_two_way(world, links_house, 'Links House Exit', player)
+        if links_house in bomb_shop_doors:
+            bomb_shop_doors.remove(links_house)
+        if links_house in blacksmith_doors:
+            blacksmith_doors.remove(links_house)
 
         # at this point only Light World death mountain entrances remain
         # place old man, has limited options
@@ -154,6 +165,16 @@ def link_entrances(world, player):
         bomb_shop_doors = list(Bomb_Shop_Single_Cave_Doors + Bomb_Shop_Multi_Cave_Doors)
         blacksmith_doors = list(Blacksmith_Single_Cave_Doors + Blacksmith_Multi_Cave_Doors)
         door_targets = list(Single_Cave_Targets)
+
+        # place links house
+        if world.mode[player] == 'standard' or not world.shufflelinks[player]:
+            links_house = 'Links House'
+        else:
+            links_house_doors = [i for i in lw_entrances if i not in Isolated_LH_Doors_Open]
+            links_house = random.choice(links_house_doors)
+        connect_two_way(world, links_house, 'Links House Exit', player)
+        if links_house in lw_entrances:
+            lw_entrances.remove(links_house)
 
         # tavern back door cannot be shuffled yet
         connect_doors(world, ['Tavern North'], ['Tavern'], player)
@@ -299,7 +320,18 @@ def link_entrances(world, player):
         else:
             dw_entrances.append('Ganons Tower')
             caves.append('Ganons Tower Exit')
-        
+
+        # place links house
+        if world.mode[player] == 'standard' or not world.shufflelinks[player]:
+            links_house = 'Links House'
+        else:
+            links_house_doors = [i for i in lw_entrances + lw_must_exits if i not in Isolated_LH_Doors_Open]
+            links_house = random.choice(links_house_doors)
+        connect_two_way(world, links_house, 'Links House Exit', player)
+        if links_house in lw_entrances:
+            lw_entrances.remove(links_house)
+        if links_house in lw_must_exits:
+            lw_must_exits.remove(links_house)
 
         # we randomize which world requirements we fulfill first so we get better dungeon distribution
         #we also places the Old Man House at this time to make sure he can be connected to the desert one way
@@ -405,6 +437,22 @@ def link_entrances(world, player):
         else:
             entrances.append('Ganons Tower')
             caves.append('Ganons Tower Exit')
+
+        # place links house
+        if world.mode[player] == 'standard' or not world.shufflelinks[player]:
+            links_house = 'Links House'
+        else:
+            links_house_doors = [i for i in entrances + must_exits if i not in Isolated_LH_Doors_Open]
+            if world.doorShuffle[player] == 'crossed' and world.intensity[player] >= 3:
+                exclusions = DW_Entrances + DW_Dungeon_Entrances + DW_Single_Cave_Doors\
+                          + DW_Entrances_Must_Exit + DW_Dungeon_Entrances_Must_Exit + ['Ganons Tower']
+                links_house_doors = [i for i in links_house_doors if i not in exclusions]
+            links_house = random.choice(list(links_house_doors))
+        connect_two_way(world, links_house, 'Links House Exit', player)
+        if links_house in entrances:
+            entrances.remove(links_house)
+        elif links_house in must_exits:
+            must_exits.remove(links_house)
 
         #place must-exit caves 
         connect_mandatory_exits(world, entrances, caves, must_exits, player)
@@ -775,12 +823,12 @@ def link_entrances(world, player):
     elif world.shuffle[player] == 'insanity':
         # beware ye who enter here
 
-        entrances = LW_Entrances + LW_Dungeon_Entrances + DW_Entrances + DW_Dungeon_Entrances + Old_Man_Entrances + ['Skull Woods Second Section Door (East)', 'Skull Woods First Section Door', 'Kakariko Well Cave', 'Bat Cave Cave', 'North Fairy Cave', 'Sanctuary', 'Lost Woods Hideout Stump', 'Lumberjack Tree Cave']
         entrances_must_exits = DW_Entrances_Must_Exit + DW_Dungeon_Entrances_Must_Exit + LW_Dungeon_Entrances_Must_Exit + ['Skull Woods Second Section Door (West)']
 
         doors = LW_Entrances + LW_Dungeon_Entrances + LW_Dungeon_Entrances_Must_Exit + ['Kakariko Well Cave', 'Bat Cave Cave', 'North Fairy Cave', 'Sanctuary', 'Lost Woods Hideout Stump', 'Lumberjack Tree Cave'] + Old_Man_Entrances +\
                 DW_Entrances + DW_Dungeon_Entrances + DW_Entrances_Must_Exit + DW_Dungeon_Entrances_Must_Exit + ['Skull Woods First Section Door', 'Skull Woods Second Section Door (East)', 'Skull Woods Second Section Door (West)'] +\
                 LW_Single_Cave_Doors + DW_Single_Cave_Doors
+        exit_pool = list(doors)
 
         # TODO: there are other possible entrances we could support here by way of exiting from a connector,
         # and rentering to find bomb shop. However appended list here is all those that we currently have
@@ -818,7 +866,7 @@ def link_entrances(world, player):
             hole_entrances.append('Hyrule Castle Secret Entrance Drop')
             hole_targets.append('Hyrule Castle Secret Entrance')
             doors.append('Hyrule Castle Secret Entrance Stairs')
-            entrances.append('Hyrule Castle Secret Entrance Stairs')
+            exit_pool.append('Hyrule Castle Secret Entrance Stairs')
             caves.append('Hyrule Castle Secret Entrance Exit')
 
         if not world.shuffle_ganon:
@@ -826,16 +874,16 @@ def link_entrances(world, player):
             connect_two_way(world, 'Pyramid Entrance', 'Pyramid Exit', player)
             connect_entrance(world, 'Pyramid Hole', 'Pyramid', player)
         else:
-            entrances.append('Ganons Tower')
             caves.extend(['Ganons Tower Exit', 'Pyramid Exit'])
             hole_entrances.append('Pyramid Hole')
             hole_targets.append('Pyramid')
             entrances_must_exits.append('Pyramid Entrance')
+            exit_pool.extend(['Ganons Tower', 'Pyramid Entrance'])
             doors.extend(['Ganons Tower', 'Pyramid Entrance'])
 
         random.shuffle(hole_entrances)
         random.shuffle(hole_targets)
-        random.shuffle(entrances)
+        random.shuffle(exit_pool)
 
         # fill up holes
         for hole in hole_entrances:
@@ -849,8 +897,22 @@ def link_entrances(world, player):
             caves.append(('Hyrule Castle Exit (West)', 'Hyrule Castle Exit (East)'))
         else:
             doors.append('Hyrule Castle Entrance (South)')
-            entrances.append('Hyrule Castle Entrance (South)')
+            exit_pool.append('Hyrule Castle Entrance (South)')
             caves.append(('Hyrule Castle Exit (South)', 'Hyrule Castle Exit (West)', 'Hyrule Castle Exit (East)'))
+
+        # place links house
+        if world.mode[player] == 'standard' or not world.shufflelinks[player]:
+            links_house = 'Links House'
+        else:
+            links_house_doors = [i for i in doors if i not in Isolated_LH_Doors_Open]
+            if world.doorShuffle[player] == 'crossed' and world.intensity[player] >= 3:
+                exclusions = DW_Entrances + DW_Dungeon_Entrances + DW_Single_Cave_Doors \
+                             + DW_Entrances_Must_Exit + DW_Dungeon_Entrances_Must_Exit + ['Ganons Tower']
+                links_house_doors = [i for i in links_house_doors if i not in exclusions]
+            links_house = random.choice(links_house_doors)
+        connect_two_way(world, links_house, 'Links House Exit', player)
+        exit_pool.remove(links_house)
+        doors.remove(links_house)
 
         # now let's deal with mandatory reachable stuff
         def extract_reachable_exit(cavelist):
@@ -869,7 +931,7 @@ def link_entrances(world, player):
             cavelist.remove(candidate)
             return candidate
 
-        def connect_reachable_exit(entrance, caves, doors):
+        def connect_reachable_exit(entrance, caves, doors, exit_pool):
             cave = extract_reachable_exit(caves)
 
             exit = cave[-1]
@@ -877,18 +939,19 @@ def link_entrances(world, player):
             connect_exit(world, exit, entrance, player)
             connect_entrance(world, doors.pop(), exit, player)
             # rest of cave now is forced to be in this world
+            exit_pool.remove(entrance)
             caves.append(cave)
 
         # connect mandatory exits
         for entrance in entrances_must_exits:
-            connect_reachable_exit(entrance, caves, doors)
+            connect_reachable_exit(entrance, caves, doors, exit_pool)
 
         # place old man, has limited options
         # exit has to come from specific set of doors, the entrance is free to move about
-        old_man_entrances = [entrance for entrance in old_man_entrances if entrance in entrances]
+        old_man_entrances = [entrance for entrance in old_man_entrances if entrance in exit_pool]
         random.shuffle(old_man_entrances)
         old_man_exit = old_man_entrances.pop()
-        entrances.remove(old_man_exit)
+        exit_pool.remove(old_man_exit)
 
         connect_exit(world, 'Old Man Cave Exit (East)', old_man_exit, player)
         connect_entrance(world, doors.pop(), 'Old Man Cave Exit (East)', player)
@@ -900,6 +963,7 @@ def link_entrances(world, player):
         blacksmith_hut = blacksmith_doors.pop()
         connect_entrance(world, blacksmith_hut, 'Blacksmiths Hut', player)
         doors.remove(blacksmith_hut)
+        exit_pool.remove(blacksmith_hut)
 
         # place dam and pyramid fairy, have limited options
         bomb_shop_doors = [door for door in bomb_shop_doors if door in doors]
@@ -907,6 +971,7 @@ def link_entrances(world, player):
         bomb_shop = bomb_shop_doors.pop()
         connect_entrance(world, bomb_shop, 'Big Bomb Shop', player)
         doors.remove(bomb_shop)
+        exit_pool.remove(bomb_shop)
 
         # handle remaining caves
         for cave in caves:
@@ -914,7 +979,7 @@ def link_entrances(world, player):
                 cave = (cave,)
 
             for exit in cave:
-                connect_exit(world, exit, entrances.pop(), player)
+                connect_exit(world, exit, exit_pool.pop(), player)
                 connect_entrance(world, doors.pop(), exit, player)
 
         # place remaining doors
@@ -1196,8 +1261,11 @@ def link_inverted_entrances(world, player):
             connect_two_way(world, entrance2, exit2, player)
         
         # place links house
-        links_house_doors = [i for i in bomb_shop_doors + blacksmith_doors if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
-        links_house = random.choice(list(links_house_doors))
+        if not world.shufflelinks[player]:
+            links_house = 'Inverted Links House'
+        else:
+            links_house_doors = [i for i in DW_Single_Cave_Doors if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
+            links_house = random.choice(links_house_doors)
         connect_two_way(world, links_house, 'Inverted Links House Exit', player)
         if links_house in bomb_shop_doors:
             bomb_shop_doors.remove(links_house)
@@ -1273,15 +1341,14 @@ def link_inverted_entrances(world, player):
         door_targets = list(Inverted_Single_Cave_Targets)
 
         # place links house
-        links_house_doors = [i for i in lw_entrances + dw_entrances + lw_must_exits if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
-        links_house = random.choice(list(links_house_doors))
+        if not world.shufflelinks[player]:
+            links_house = 'Inverted Links House'
+        else:
+            links_house_doors = [i for i in dw_entrances if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
+            links_house = random.choice(links_house_doors)
         connect_two_way(world, links_house, 'Inverted Links House Exit', player)
-        if links_house in lw_entrances:
-            lw_entrances.remove(links_house)
-        elif links_house in dw_entrances:
+        if links_house in dw_entrances:
             dw_entrances.remove(links_house)
-        elif links_house in lw_must_exits:
-            lw_must_exits.remove(links_house)
 
         # place dark sanc
         sanc_doors = [door for door in Inverted_Dark_Sanctuary_Doors if door in dw_entrances]
@@ -1403,15 +1470,14 @@ def link_inverted_entrances(world, player):
         caves.remove('Inverted Agahnims Tower Exit')
         
         # place links house
-        links_house_doors = [i for i in lw_entrances + dw_entrances + lw_must_exits if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
-        links_house = random.choice(list(links_house_doors))
+        if not world.shufflelinks[player]:
+            links_house = 'Inverted Links House'
+        else:
+            links_house_doors = [i for i in dw_entrances if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
+            links_house = random.choice(links_house_doors)
         connect_two_way(world, links_house, 'Inverted Links House Exit', player)
-        if links_house in lw_entrances:
-            lw_entrances.remove(links_house)
         if links_house in dw_entrances:
             dw_entrances.remove(links_house)
-        if links_house in lw_must_exits:
-            lw_must_exits.remove(links_house)
 
         # place dark sanc
         sanc_doors = [door for door in Inverted_Dark_Sanctuary_Doors if door in dw_entrances]
@@ -1525,7 +1591,7 @@ def link_inverted_entrances(world, player):
             hc_ledge_entrances = ['Hyrule Castle Entrance (West)', 'Hyrule Castle Entrance (East)', 'Inverted Ganons Tower']
 
         # shuffle aga door. if it's on hc ledge, then one other hc ledge door has to be must_exit
-        aga_door = random.choice(list(entrances))
+        aga_door = random.choice(entrances)
         
         if aga_door in hc_ledge_entrances:
             hc_ledge_entrances.remove(aga_door)
@@ -1539,10 +1605,12 @@ def link_inverted_entrances(world, player):
         connect_two_way(world, aga_door, 'Inverted Agahnims Tower Exit', player)
         caves.remove('Inverted Agahnims Tower Exit')
 
-
         # place links house
-        links_house_doors = [i for i in entrances + must_exits if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
-        links_house = random.choice(list(links_house_doors))
+        if not world.shufflelinks[player]:
+            links_house = 'Inverted Links House'
+        else:
+            links_house_doors = [i for i in entrances + must_exits if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
+            links_house = random.choice(links_house_doors)
         connect_two_way(world, links_house, 'Inverted Links House Exit', player)
         if links_house in entrances:
             entrances.remove(links_house)
@@ -1605,20 +1673,18 @@ def link_inverted_entrances(world, player):
     elif world.shuffle[player] == 'insanity':
         # beware ye who enter here
 
-        entrances = Inverted_LW_Entrances + Inverted_LW_Dungeon_Entrances + Inverted_DW_Entrances + Inverted_DW_Dungeon_Entrances + Inverted_Old_Man_Entrances + Old_Man_Entrances + ['Skull Woods Second Section Door (East)', 'Skull Woods Second Section Door (West)', 'Skull Woods First Section Door', 'Kakariko Well Cave', 'Bat Cave Cave', 'North Fairy Cave', 'Sanctuary', 'Lost Woods Hideout Stump', 'Lumberjack Tree Cave', 'Hyrule Castle Entrance (South)']
         entrances_must_exits = Inverted_LW_Entrances_Must_Exit + Inverted_LW_Dungeon_Entrances_Must_Exit
 
         doors = Inverted_LW_Entrances + Inverted_LW_Dungeon_Entrances + Inverted_LW_Entrances_Must_Exit + Inverted_LW_Dungeon_Entrances_Must_Exit + ['Kakariko Well Cave', 'Bat Cave Cave', 'North Fairy Cave', 'Sanctuary', 'Lost Woods Hideout Stump', 'Lumberjack Tree Cave', 'Hyrule Castle Secret Entrance Stairs'] + Inverted_Old_Man_Entrances +\
                 Inverted_DW_Entrances + Inverted_DW_Dungeon_Entrances + ['Skull Woods First Section Door', 'Skull Woods Second Section Door (East)', 'Skull Woods Second Section Door (West)'] +\
                 Inverted_LW_Single_Cave_Doors + Inverted_DW_Single_Cave_Doors + ['Desert Palace Entrance (West)', 'Desert Palace Entrance (North)']
+        exit_pool = list(doors)
 
         # randomize which desert ledge door is a must-exit
         if random.randint(0, 1) == 0:
             entrances_must_exits.append('Desert Palace Entrance (North)')
-            entrances.append('Desert Palace Entrance (West)')
         else:
             entrances_must_exits.append('Desert Palace Entrance (West)')
-            entrances.append('Desert Palace Entrance (North)')
 
         # TODO: there are other possible entrances we could support here by way of exiting from a connector,
         # and rentering to find bomb shop. However appended list here is all those that we currently have
@@ -1648,7 +1714,6 @@ def link_inverted_entrances(world, player):
 
         hole_entrances.append('Hyrule Castle Secret Entrance Drop')
         hole_targets.append('Hyrule Castle Secret Entrance')
-        entrances.append('Hyrule Castle Secret Entrance Stairs')
         caves.append('Hyrule Castle Secret Entrance Exit')
 
         if not world.shuffle_ganon:
@@ -1656,15 +1721,15 @@ def link_inverted_entrances(world, player):
             connect_two_way(world, 'Inverted Pyramid Entrance', 'Pyramid Exit', player)
             connect_entrance(world, 'Inverted Pyramid Hole', 'Pyramid', player)
         else:
-            entrances.append('Inverted Ganons Tower')
             caves.extend(['Inverted Ganons Tower Exit', 'Pyramid Exit'])
             hole_entrances.append('Inverted Pyramid Hole')
             hole_targets.append('Pyramid')
             doors.extend(['Inverted Ganons Tower', 'Inverted Pyramid Entrance'])
+            exit_pool.extend(['Inverted Ganons Tower', 'Inverted Pyramid Entrance'])
 
         random.shuffle(hole_entrances)
         random.shuffle(hole_targets)
-        random.shuffle(entrances)
+        random.shuffle(exit_pool)
 
         # fill up holes
         for hole in hole_entrances:
@@ -1674,18 +1739,18 @@ def link_inverted_entrances(world, player):
         caves.append(('Hyrule Castle Exit (South)', 'Hyrule Castle Exit (West)', 'Hyrule Castle Exit (East)'))
         
         # place links house and dark sanc
-        links_house_doors = [i for i in entrances + entrances_must_exits if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
-        links_house = random.choice(list(links_house_doors))
+        if not world.shufflelinks[player]:
+            links_house = 'Inverted Links House'
+        else:
+            links_house_doors = [i for i in doors if i not in Inverted_Dark_Sanctuary_Doors + Isolated_LH_Doors]
+            links_house = random.choice(links_house_doors)
         connect_two_way(world, links_house, 'Inverted Links House Exit', player)
-        if links_house in entrances:
-            entrances.remove(links_house)
-        elif links_house in entrances_must_exits:
-            entrances_must_exits.remove(links_house)
         doors.remove(links_house)
+        exit_pool.remove(links_house)
 
-        sanc_doors = [door for door in Inverted_Dark_Sanctuary_Doors if door in entrances]
+        sanc_doors = [door for door in Inverted_Dark_Sanctuary_Doors if door in exit_pool]
         sanc_door = random.choice(sanc_doors)
-        entrances.remove(sanc_door)
+        exit_pool.remove(sanc_door)
         doors.remove(sanc_door)
         connect_entrance(world, sanc_door, 'Inverted Dark Sanctuary', player)
         world.get_entrance('Inverted Dark Sanctuary Exit', player).connect(world.get_entrance(sanc_door, player).parent_region)
@@ -1707,7 +1772,7 @@ def link_inverted_entrances(world, player):
             cavelist.remove(candidate)
             return candidate
 
-        def connect_reachable_exit(entrance, caves, doors):
+        def connect_reachable_exit(entrance, caves, doors, exit_pool):
             cave = extract_reachable_exit(caves)
 
             exit = cave[-1]
@@ -1715,18 +1780,19 @@ def link_inverted_entrances(world, player):
             connect_exit(world, exit, entrance, player)
             connect_entrance(world, doors.pop(), exit, player)
             # rest of cave now is forced to be in this world
+            exit_pool.remove(entrance)
             caves.append(cave)
 
         # connect mandatory exits
         for entrance in entrances_must_exits:
-            connect_reachable_exit(entrance, caves, doors)
+            connect_reachable_exit(entrance, caves, doors, exit_pool)
 
         # place old man, has limited options
         # exit has to come from specific set of doors, the entrance is free to move about
-        old_man_entrances = [entrance for entrance in old_man_entrances if entrance in entrances]
+        old_man_entrances = [entrance for entrance in old_man_entrances if entrance in exit_pool]
         random.shuffle(old_man_entrances)
         old_man_exit = old_man_entrances.pop()
-        entrances.remove(old_man_exit)
+        exit_pool.remove(old_man_exit)
 
         connect_exit(world, 'Old Man Cave Exit (East)', old_man_exit, player)
         connect_entrance(world, doors.pop(), 'Old Man Cave Exit (East)', player)
@@ -1738,6 +1804,7 @@ def link_inverted_entrances(world, player):
         blacksmith_hut = blacksmith_doors.pop()
         connect_entrance(world, blacksmith_hut, 'Blacksmiths Hut', player)
         doors.remove(blacksmith_hut)
+        exit_pool.remove(blacksmith_hut)
 
         # place dam and pyramid fairy, have limited options
         bomb_shop_doors = [door for door in bomb_shop_doors if door in doors]
@@ -1745,6 +1812,7 @@ def link_inverted_entrances(world, player):
         bomb_shop = bomb_shop_doors.pop()
         connect_entrance(world, bomb_shop, 'Inverted Big Bomb Shop', player)
         doors.remove(bomb_shop)
+        exit_pool.remove(bomb_shop)
 
         # handle remaining caves
         for cave in caves:
@@ -1752,7 +1820,7 @@ def link_inverted_entrances(world, player):
                 cave = (cave,)
 
             for exit in cave:
-                connect_exit(world, exit, entrances.pop(), player)
+                connect_exit(world, exit, exit_pool.pop(), player)
                 connect_entrance(world, doors.pop(), exit, player)
 
         # place remaining doors
@@ -2436,7 +2504,19 @@ LW_Single_Cave_Doors = ['Blinds Hideout',
                         'Kings Grave',
                         'Bonk Fairy (Light)',
                         'Hookshot Fairy',
-                        'Mimic Cave']
+                        'Mimic Cave',
+                        'Links House']
+
+Isolated_LH_Doors_Open = ['Mimic Cave',
+                          'Kings Grave',
+                          'Waterfall of Wishing',
+                          'Desert Palace Entrance (South)',
+                          'Desert Palace Entrance (North)',
+                          'Capacity Upgrade',
+                          'Ice Palace',
+                          'Skull Woods Final Section',
+                          'Dark World Hammer Peg Cave',
+                          'Turtle Rock Isolated Ledge Entrance']
 
 DW_Single_Cave_Doors = ['Bonk Fairy (Dark)',
                         'Dark Sanctuary Hint',
@@ -2724,7 +2804,7 @@ Inverted_Bomb_Shop_Multi_Cave_Doors = ['Hyrule Castle Entrance (South)',
 
 Inverted_Blacksmith_Multi_Cave_Doors = Blacksmith_Multi_Cave_Doors  # same as non-inverted
 
-Inverted_LW_Single_Cave_Doors = LW_Single_Cave_Doors + ['Inverted Big Bomb Shop']
+Inverted_LW_Single_Cave_Doors = [x for x in LW_Single_Cave_Doors if x != 'Links House'] + ['Inverted Big Bomb Shop']
 
 Inverted_DW_Single_Cave_Doors = ['Bonk Fairy (Dark)',
                                  'Inverted Dark Sanctuary',
@@ -3215,7 +3295,9 @@ inverted_mandatory_connections = [('Links House S&Q', 'Inverted Links House'),
                                   ('Bomb Hut Outer Bushes', 'Bomb Hut Area'),
                                   ('Bomb Hut Mirror Spot', 'West Dark World')]
 # non-shuffled entrance links
-default_connections = [('Waterfall of Wishing', 'Waterfall of Wishing'),
+default_connections = [('Links House', 'Links House'),
+                       ('Links House Exit', 'Light World'),
+                       ('Waterfall of Wishing', 'Waterfall of Wishing'),
                        ("Blinds Hideout", "Blinds Hideout"),
                        ('Dam', 'Dam'),
                        ('Lumberjack House', 'Lumberjack House'),